CDH-56538: Sentry code to read new properties for list of sentry servers
As discussed in CDH-56195, we would want to extend 5.13 sentry to be
able to read new properties for getting the list of sentry servers. This
will ensure config compatibility while using older code base of Sentry.

Thinking of making this as a CDH5 only commit. Once we move to
Sentry-2.0 during C6 timeframe, it will continue to be using original
properties and we don't have to worry about maintaining compatibility
with older releases.

Change-Id: I173f6256d46fa729c84591e84a2562b4102ac795
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/24396
Tested-by: Jenkins User
Reviewed-by: Na Li <lina.li@cloudera.com>