CLOUDERA-BUILD Use correct branch name
The branching script defaults to performing a raw string replacement of
the maven branch name. In this case, the replacement was:

6\.x-SNAPSHOT to 6.0.x-SNAPSHOT

The strings below didn't match, which caused unit tests in the cdh6.0.x
branch to fail. An alternative here is to create a cloudera/branch.sh
script, as described here.

http://github.mtv.cloudera.com/CDH/cdh/blob/cdh6.x/README_cauldron.md#branching-and-branch-names

For now, this commit should fix unit tests.

Change-Id: I38fe577fb2e1866ea3409840dac999aa814d4ca6