CDH-69573: Port partially SENTRY-711 to list user privileges for authorization
==C5_FEATURE_FGP==

SENTRY-711 implemented grant user to a role as a new feature and added a new parameter in listPrivilegesForProvider() to request privileges for users in a role.
However, this method is used upstream to return user privileges as well, not just user to role privileges.

This patch covers a partial backport of SENTRY-711 to add the new API on the Sentry clients, cache and backend interfaces.

Change-Id: I9d9774f676bc3989b7ca1f454ded964828ee3a0c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/35581
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Kalyan Kumar Kalvagadda <kkalyan@cloudera.com>